fix warning of data type

+++ capture_thread.c
@@ -101,7 +101,7 @@ void *live_thread_pfring(void *arg) {
 			delay = statis.tv_cur.tv_sec - statis.tv_old.tv_sec;
 			if(delay >= 60) {
 				pfring_stats(pfhdl, &stat);
-				printf("Live NIC %s, %s recv:%lu drop:%lu, %.2f pps\n",dev,ringname,stat.recv,stat.drop,(statis.pkt_cur-statis.pkt_old)/(float)delay);
+				printf("Live NIC %s, %s recv:%lu drop:%lu, %.2f pps\n",dev,ringname,(unsigned long)stat.recv,(unsigned long)stat.drop,(statis.pkt_cur-statis.pkt_old)/(float)delay);
 				statis.pkt_old = statis.pkt_cur;
 				statis.tv_old = statis.tv_cur;
 			}
